pyproject.toml
[tool.poetry]
name = "tasknote"
version = "1.0.0"
description = ""
authors = ["Kauã Mateus"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-cors-headers = "^4.4.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
django-grappelli = "^4.0.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
